Rename `logging` to `loggers`, to avoid collision with Python's module
