+++ kopf/cli.py
@@ -5,7 +5,7 @@ from typing import Any, Callable, List, Optional
 
 import click
 
-from kopf.engines import logging, peering
+from kopf.engines import loggers, peering
 from kopf.reactor import registries, running
 from kopf.structs import configuration, credentials, primitives
 from kopf.utilities import loaders
@@ -28,7 +28,7 @@ def logging_options(fn: Callable[..., Any]) -> Callable[..., Any]:
     @click.option('-q', '--quiet', is_flag=True)
     @functools.wraps(fn)  # to preserve other opts/args
     def wrapper(verbose: bool, quiet: bool, debug: bool, *args: Any, **kwargs: Any) -> Any:
-        logging.configure(debug=debug, verbose=verbose, quiet=quiet)
+        loggers.configure(debug=debug, verbose=verbose, quiet=quiet)
         return fn(*args, **kwargs)
 
     return wrapper
