+++ tests/conftest.py
@@ -16,7 +16,7 @@ import pytest_mock
 
 import kopf
 from kopf.clients.auth import APIContext
-from kopf.engines.logging import ObjectPrefixingFormatter, configure
+from kopf.engines.loggers import ObjectPrefixingFormatter, configure
 from kopf.engines.posting import settings_var
 from kopf.structs.configuration import OperatorSettings
 from kopf.structs.credentials import ConnectionInfo, Vault, VaultKey
